requirements.txt
# Core Django
Django>=4.2.0,<5.0

# Django REST Framework
djangorestframework>=3.14.0

# Django Filters for DRF
django-filter>=23.2

# Widget Tweaks for form customization
django-widget-tweaks>=1.4.12

# Python-dotenv for environment variables
python-dotenv>=1.0.0

# Pillow for image handling (if needed for character/deity images)
Pillow>=10.0.0

# Gunicorn for production server
gunicorn>=21.2.0

# Psycopg2 for PostgreSQL database (Railway's preferred database)
psycopg2-binary>=2.9.7

# dj-database-url for parsing database URLs
dj-database-url>=2.1.0